fix: use ref for increment in Counter component
The timer in the Counter component was not updating with the latest increment value because it was using the initial value of the increment state variable. This commit fixes the problem by using a ref to store the current value of the increment state variable and updating the ref whenever the increment state variable is updated.